메인피드에서 포스트 컴포넌트화 하고 포스트의 이미지 슬라이드 작업
메인피드에서 포스트 컴포넌트화 하고 포스트의 이미지 슬라이드 작업
